Fix tokenizer deadlock
The previous tokenizer implementation can only handle a small amount of
input. If given a large amount of input, the main process and the
subprocess will become deadlocked. The main process will be unable to
write any more to the stdin buffer, and the subprocess will be unable to
write any more to the stdout buffer.

The solution is to read the subprocess' stdout from a separate thread so
that the subprocess' stdout buffer can be emptied as the main process
writess to the subprocess' stdin concurrently. This reader thread needs
to be created using `read` prior to calling `tokenize` now.